move PubkeyResult to pgp-key
