+++ extension/chrome/elements/compose-modules/formatters/encrypted-mail-msg-formatter.ts
@@ -6,14 +6,14 @@ import { Backend, FcUuidAuth } from '../../../../js/common/api/backend.js';
 import { BaseMailFormatter } from './base-mail-formatter.js';
 import { ComposerResetBtnTrigger } from '../compose-err-module.js';
 import { Mime, SendableMsgBody } from '../../../../js/common/core/mime.js';
-import { NewMsgData, PubkeyResult } from '../compose-types.js';
+import { NewMsgData } from '../compose-types.js';
 import { Str, Value } from '../../../../js/common/core/common.js';
 import { ApiErr } from '../../../../js/common/api/error/api-error.js';
 import { Att } from '../../../../js/common/core/att.js';
 import { Buf } from '../../../../js/common/core/buf.js';
 import { Catch } from '../../../../js/common/platform/catch.js';
 import { Lang } from '../../../../js/common/lang.js';
-import { PgpKey } from '../../../../js/common/core/pgp-key.js';
+import { PgpKey, PubkeyResult } from '../../../../js/common/core/pgp-key.js';
 import { PgpMsg, PgpMsgMethod } from '../../../../js/common/core/pgp-msg.js';
 import { SendableMsg } from '../../../../js/common/api/email-provider/sendable-msg.js';
 import { Settings } from '../../../../js/common/settings.js';
